+++ pom.xml
@@ -79,6 +79,7 @@
     <maven-gpg-plugin.version>3.0.1</maven-gpg-plugin.version>
     <qulice-maven-plugin.version>0.22.0</qulice-maven-plugin.version>
     <oop-cop.version>0.2.6</oop-cop.version>
+    <cactoos-matchers.version>0.25</cactoos-matchers.version>
   </properties>
   <dependencies>
     <dependency>
@@ -152,6 +153,18 @@
       <artifactId>lombok</artifactId>
       <version>${lombok.version}</version>
       <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.llorllale</groupId>
+      <artifactId>cactoos-matchers</artifactId>
+      <version>${cactoos-matchers.version}</version>
+      <scope>test</scope>
+      <exclusions>
+        <exclusion>
+          <groupId>org.cactoos</groupId>
+          <artifactId>cactoos</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
   </dependencies>
   <build>
