+++ src/main/java/ru/l3r8y/checks/ErSuffixCheck.java
@@ -21,68 +21,53 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package ru.l3r8y.rule;
+package ru.l3r8y.checks;
 
 import java.util.Collection;
 import lombok.RequiredArgsConstructor;
+import ru.l3r8y.Check;
 import ru.l3r8y.ClassName;
 import ru.l3r8y.Complaint;
-import ru.l3r8y.Rule;
 import ru.l3r8y.complaint.ClassifiedComplaint;
 import ru.l3r8y.complaint.LinkedComplaint;
 import ru.l3r8y.complaint.WrongClassNaming;
 
 /**
- * Check for Long class name.
+ * It checks if class is -er named.
  *
- * @since 0.2.0
+ * @since 0.1.6
  */
 @RequiredArgsConstructor
-public final class LongClassNameCheck implements Rule {
+public final class ErSuffixCheck implements Check {
 
     /**
-     * Class name.
+     * The name of class to check.
      */
     private final ClassName name;
 
-    /**
-     * Length to pass.
-     */
-    private final int fine;
-
     @Override
     public Collection<Complaint> complaints() {
-        return new ConditionRule(
-            this::longerThanOk,
+        return new ConditionCheck(
+            this::isEndsWithEr,
             new ClassifiedComplaint(
                 new LinkedComplaint(
                     new WrongClassNaming(
                         this.name,
-                        String.format(
-                            "class name is more than %s, consider more simple name",
-                            this.fine
-                        )
+                        "class ends with '-er' suffix, it's prohibited"
                     ),
-                    "https://www.yegor256.com/2015/01/12/compound-name-is-code-smell.html"
+                    "https://www.yegor256.com/2015/03/09/objects-end-with-er.html"
                 ),
                 this.getClass()
             )
         ).complaints();
     }
 
-    /*
-     * @todo #81 Introduce new class from #longerThanOk private method.
-     *  we should introduce new reusable class from private method
-     *  #longerThanOk.
-     *  Don't forget to remove this puzzle.
-     */
-
     /**
-     * Is name ok?
+     * Is ends with -er.
      *
-     * @return True if longer
+     * @return True if ends with '-er'.
      */
-    private boolean longerThanOk() {
-        return this.name.value().length() > this.fine;
+    private boolean isEndsWithEr() {
+        return this.name.value().endsWith("er");
     }
 }
